functioning mobile view
